source deps should be installed before apt update
In the example of ubuntu cloud archive, you need to
specify that 'ubuntu-cloud-keyring' is a required package.

In this case, the package needs to be installed before
apt-get-update is run, or the source cannot be used
(b/c its key will not be installed). I am assuming that
is generally applicable (that source dependencies need to
be installed before the apt cache is updated)

Although I added tests, the spec tests do not work b/c another
patch already broke tests.